Added "possessive" attribute
Modified various types to include a "possessive" type, set to "its", "his, "her", "their" or "your". Also modified the editor files so this appears on the object tab, under the "gender" and "article".

Also added "me" to the list of alternative aliases for the player object, so X ME will still work if the author gives the player object a proper name.