Improved with web users, re-ordered, etc,
Added a page about status attributes. This was previously mentioned in the turn scripts section, but it was hidden away, it was after the page on changing the player object, which talks about them, and was restricted to desktop. The new page goes before the changing the player page, and gives a method for web users (albeit not a simple method).

Moved templates to the elite section.

Split "text formatting" into "multimedia" and "changing the user experience", and enhanced the latter

Re-arranged the advanced section, so scripting stuff is first, then fancy looks, then steps towards release.

Added a key to unlocking an exit, and put it after locked containers.

Rewritten the JavaScript page, though it might need more work yet!

Some other minor changes, including work-arounds with web users where possible.