Tutorial reorganised
The tutorial has been reduced to essentially just the first two sections, which have been combined into a single unit. Hopefully all authors will be able to complete the full tutorial, and then can select what they need from the rest of the documentation.
Note that pages that are dropped from the tutorial are still in the tutorial folder, the only changes I have made is to delete the links at the end of the pages from those pages and updated the links where necessary for pages that are kept (besides index which has been edited significantly).