Web users should add synonym verbs separately
Every couple of months we get someone who is using the web version and has added a verb that messes up their game. Quest already checks that they are not adding "use" or "look at", but not if that is part of a list, such as "flick; use" as cropped up recently. Ideally the C# code would check this, but as an alternative, these pages recommend only adding one verb at a time.