Easy-access version 5.7
This is version 5.7 beta. It is not a release version. It is packaged with an installer, but it will issue a warning when you install saying it is from an untrusted source, as I have yet to get a certificate. Also, the copyright notices need updating. However, the software itself should be good to go.