Allow initialise to be edited in the GUI
An earlier change added the option to have an _initialise_ attribute on any object; a script that will run at the start of the game. This upload makes that a feature, so it can be ticked on the Feature tab, to give a new a new tab. This will allow users of the web version to use it, which in some way mitigates their not being able to add attribute directly - now they can add attributes for an object on the object itself (for rooms, you can use the before first enter script to do the same thing).