Added TestRemove function
This is the opposite of TestGarment, which tests if an item can be put on.  TestRemove checks if if can be taken off.
Also, successful commands use `P` function to print. This just uses msg, so there is no effect, but I have a plan ultimately (bt not in this version) to have all successful commands use `P`, allowing successful and unsuccessful to be differentiated (say with unsuccessful in grey). So far, this is just in CoreWearables as this is new.