either and select split on |
Found an issue with the "either" text processor directive that means other directives cannot be nested in it. This is solved by splitting the two text sections by a vertical bar instead of a colon.
Found the same issue in the "select" (which was in earlier versions), so modified that so that if there are vertical bars in the text, his will split on them, allowing nesting, and if not it will split on colons as it always has.