Now rounds to zero
A bug was repored during beta-testing of 5.7 from Dr Agon:

I ticked the box for 'set page opacity' and pressed the 'up' button to '1', then the down button. 0.4, 0.3, 0.2, 0.1, 1.38777878078145E-16
Failed to load game. Invalid number specified

This was solved by modifying ctlNumber_ValueChanged so that a value of less than 0.0001 is set to zero. Note that as the control can potentially be used for any range of numbers this could impact on its later use if the range is small enough. However, currently the control is only used for page opacity, so this is fine.

Possibly a better solution would be to do opacity as a percentage so it is an int, and just remove this control altogether.