ShowMenu will now accept object lists
There is a facility to give objects their own hyperlink colour. Dcoder pointed out that disambiguation menus should use this too. This was tricky as ShowMenu only took string lists or string dictionaries, so the information about the object does not get to where it is needed.
I therefore changed ShowMenu to accept any list (or a string dictionary), and if an entry is an object, to use the object's hyperlink colour.
In CoreCommands, the USE and GIVE commands were modified to send the object list, rather than converting the list to a string list. Similarly in CoreParser, ResolveNameFromList and defaultverb now use the object list.
Some debugging messages in CoreParser that were commented out have been deleted.
ListCompact is a new function that takes a list, and returns it with all nulls and repeats removed.
GenerateMenuChoices is now has a comment that it is deprecated, as it is no longer used.