Various changes...
CoreParser.aslx, English.aslx - Support for AGAIN/G (from K.V.).

Core.aslx - Added JS.getDivCount () to the initialisation script to set _divCount (thanks to K.V.), and a line to set the location when loading from a save.
https://github.com/textadventures/quest/issues/917
https://github.com/textadventures/quest/issues/908

CoreParser.aslx, Core.aslx - Moved FinishTurn from the former to the latter as it is not part of the parser process

CoreParser.aslx, CoreTypesw.aslx - Moved defaultverb from the former to the latter as it is a type

CoreParser.aslx - Added lots of options to GetScope (plus GetScoping command to support it) (thanks to mrangel)

CoreParser.aslx - Rearranged the order of the functions and added lots of comments, to make it easier to understand how it works

CoreEditorVerbs.aslx - Added scope (using the command template)

CoreCommands.aslx - Added scope attributes to put/in and remove/from

CoreFunction.aslx - New CloneObjectAndInitialise that doers the same as CloneObject but also runs the initialisation script on the clone if there is one.

CoreCommands.aslx, CoreWearables.aslx, CoreParser.aslx - Modified how multiple commands (GET ALL) etc. are handled so they give a message if there is nothing there, and to use scripts rather than delegates.