Update playercore.js
Modified the clear screen function on lines 741 - 754

This keeps the ```divOutputAlign``` numbering system in order after clearing the screen and saving.

It also hides the cleared text rather than deleting it.  This will allow us to produce a game transcript.  

mrangel has already written a script that allows a transcript to be downloaded while playing online.  (It works in Chrome and Firefox right now.)  I am working on getting the desktop player to play along, too.