Option for map to respond to mouse clicks
Clicking on the map sent an ASLEvent, which authors could use, but either way cause a turn to pass. This update adds an option to the Interface tab so authors can choose to respond to mouse clicks (which will cause a turn to pass), but by default they have no effect.