Modified GoURL and GetURL to work with logs, transcripts, and SaveLoad.
Changed GoURL to open any url that does not begin with file: in the user's default browser.

Added getDefaultBrowser() to find the user's default browser for GoURL().

Modified GetURL to load a local file outside of the game's main folder, but there is a catch: The file must be in Documents\Quest Shared.