Minor feature enhancements
Status attributes improved, as per this thread:
http://textadventures.co.uk/forum/quest/topic/ppim1jbkiemsvvpv19nzkg/status-attributes

Added `QuickParam`, which creates a dictionary, and adds a single entry from the given parameters. If you want to pass parameters to a script, you can do that in a dictionary. If you just want to pass one parameter, you can now do this:
```
do (dog, "bark", QuickParam("object", cat))
```
Quest no longer uses delegates. There are still there, so if you want to use them, there is no change in that regards, and there are no plans to get rid of them. Delegates are not bad, but they are different. They were only used in a couple of places, and could be easily dropped. There is no way to use them via the editor, even with the desktop version, and I would prefer to keep Quest "mainstream", i.e., only using stuff that we can all use. 

Scope for a command can now be set to an object list attribute of the player object.