Setting game.suppressturnscripts suppresses turnscripts
Gives the creator the opportunity to suppress turnscripts for one turn, by setting game.suppressturnscripts to true. You might want to do this if a command is not understood, for example, or for HELP. Could also be used when the player is outside normal play. Or when handling mouse clicks on the map.