Multiple updates
Added spellcheck to rich text boxes

Added VERSION command

Modified take to exclude objects with 'not_all' set to true from TAKE ALL

Modified lookat to increase timesexamined attribute when successful

Added isroom boolean to the room object in all templates and to editor_room

Added AllRooms function, which returns an object lilst

Added DbgLog and DbgMsg functions

Added DictionaryAdd and DictionaryRemove functions, which check for the existence of dictionaries and keys before attempting to perform any actions.  If the dicionary does not exist, a log entry is created.  If the dictionary contains the key, the value is overwritten.