Update files
Fixed my typo in the Log function in CoreOutput.aslx (again?) and added a line for it to send the text to a UIEvent to save to the local drive.

Changed the way lists were combined in ResolveNameFromList from ```candidates = ListCompact(fullmatches + partialmatches)``` to ```candidates = ListCompact(ListCombine(fullmatches, partialmatches))``` (mainly to resolve an issue with QuestJS)

Removed SpellCheck from RichTextControl.xaml

Corrected my mistake selecting AM or PM in the log time and date function in playercore.js.