Better ALL handling
If a command has "allow_all" set to true, Quest will use GetScope when handling ALL to get the right objects, making it consistent with the rest (except scenery objects will not be taken). No multiple scripts now!
There is an issue with WEAR ALL and REMOVE ALL that they may object if the order is wrong.