Fix transcript_off_cmd and Remove unused attribute
I had the wrong JS variable in the transcript_off_command, and the transcript did not stop when disabled!  (Sorry!  The variable name is ```savingTranscript``` in JS, and the Quest attribute is ```savetranscript```.  I got confused.  Note that I can't use ```saveTranscript``` in JS because that is the name of my function.)

---
I also fixed the command patterns for each of these, replacing the English pattern with the template.  (I'd fixed this in my game, so when I copied my code to the commit, it had the English pattern copied into my code.)

---
I also removed quite a few lines of code where I was setting ```game.notarealturn``` to true.