Tweaked GET/TAKE ALL
Should now work properly with containers now. See here:
https://github.com/textadventures/quest/issues/1017

If a command uses the NotOpen template, it now uses BlockingMessage, which in turn allows a custom message on the blocking object. This means if you set your NPC to be a container, you can have Quest response "He does let you" instead of "He is not open".