Added GetDefiniteName function
Prepends "the" to the name/alias, but only if the object uses a prefix. I have added this to English.aslx as it is specific to English (and not used in the core code at all). Other language users may want to add their own version, which may depend on gender as required.