Added QuickParams function
Values can be passed to a script in a dictionary, and in the script they are just local variables, which is great. However, setting up the dictionary is a pain. QuickParams offers a quick way to do that, creating a dictionary with up to three entries in one go. The function can take 2, 4 or 6 parameters to give a dictionary with 1, 2 or 3 entries.