Fix major bug concerning transcript
This removes the majority of the code related to the transcript.

Currently, enabling a transcript crashes online games, and it stops all turns scripts from firing while it is enabled in the desktop player.

There is a totally separate, existing function in playercore.js named `showScrollback()` which handles displaying and printing the transcript all by itself. All I left intact besides that JS function are the commands and templates to view the transcript.

I altered docs\transcript.md to reflect these changes.

I also removed or disabled all the Quest and JS functions to save the transcript to a local file in the desktop version, because this slows the games down tremendously in practice.

I left the code in Player\PlayerHTML.vb to handle it if someone wanted to add the code to Quest and JS to use this feature. That way, no one would have to rebuild the project in Visual Studio just to use that functionality.

---
Double check me on the Dansk `transcript_on_cmd` change I made. The internet says it translates correctly everywhere I looked it up.