kv-4-transcript-repair
There is a game-crashing bug in the transcript code I submitted for Quest 5.8.0.

If playing online and the transcript is enabled, the game completely crashes (because `SaveTranscript` is only in desktop.js, and is therefore undefined in the web player).

If playing in the desktop and the transcript is enabled, all turn script stop working completely and if a walkthrough is being recorded, the recorded steps are the HTML being output to the game.

ALSO, there was a JS string variable named `transcriptString` which got larger and larger when adding the current text for the transcript every time text printed to the game. This obviously slowed things down. Plus, this text was also sent by `ASLEvent` to update `game.transcriptstring` each time text printed along with everything else.

Also, the updated `clearScreen()` function in playercore.js was setup to save the cleared text and hide it rather than deleting it. This also negatively effected game performance exponentially as play progressed. The JS boolean `saveClearedText` is now set to false by default.

Finally, if the VIEW TRANSCRIPT command was run while recording a walkthrough, the jQuery-UI popup would break the walkthrough during playback.

The in-game HTML transcript feature has now been removed. There is no VIEW TRANSCRIPT command anymore. There is no longer a `transcriptString` JS variable, nor is there a `game.transcriptstring` attribute.

The only way the transcript will work now is if the game is being played in the desktop player. It will save the transcript, and append it, every time text prints to the screen, to an HTML file in "Documents\Quest Transcripts", just as in 5.8.0. Except I also fixed my error in PlayerHTML.vb to make the file saves under the transcript name the player chooses when enabling the transcript.

I added a "fallback" of each function I added to Player\desktopplayer.js to WebPlayer\player.js. These functions do nothing and serve no purpose except to avoid possible errors due to being undefined.

Also added a bit of code in `HandleCommand` and `ShowMenuResponse` to add the player's command to the transcript if `game.echocommand` is `false`.

Updated the transcript doc page to reflect the changes.

I triple checked things every way I could imagine in the desktop build as well as using VS to test the web player and the mobile version.

I almost forgot: I left it where you can still call `showScrollback()` from JS and it will show the game's text in a JS popup. If `clearSavedText` is changed to `false`, it will save the cleared text in the scrollback.

The `printScrollback()` JS function is also still present. Calling it will use the player's device to print the scrollback (whether via a printer or saving it to PDF).

Again, the scroll back will not contain any cleared text unless `saveClearedText` is changed from `false` to `true`.

The scrollback JS functions are pretty much hidden, but an author could easily add code to allow their use in their game.