kv-5-modify-walkthroughrunner-and-player-vb-for-whereami
This adds IF statements to **Player\Player.vb** and **Player\WalkthroughRunner.vb** to prevent recording the step or calling "SendEvent" when the event is "WhereAmI" or the event begins with "NoEventFunc".

(I threw the "NoEventFunc*" bit in there just to give everyone wiggle room with anything they can imagine. It wouldn't hurt my feelings if this part was removed.)

In my last pull request, I added a call to `JS.whereAmI()` from `InitInterface`. Afterwards, I found an old commit from ThePix, wherein that line was removed a few years ago because it got recorded to (and called from) the walkthrough, causing odd behaviour.

https://github.com/textadventures/quest/commit/e334a8aef39cda4904484858a1fac3d8f09f2413

This fixes that issue, while allowing Quest to determine whether it is being played in the desktop player or the web player (or the mobile player) when the game loads. And this makes it so the ENABLE TRANSCRIPT command can work smoothly when first entered during play. (I'm sure others will find other uses for it, too.)