Update to TXT Transcript
An HTML transcript is a security risk because the JS could still run in it, and the HTML file format is confusing to some people.

I just learned how to save the transcript as plain text to a .txt file. (We tried to do it this way to begin with.)

---
Also moved DevMode templates from EditorEnglish.aslx and EditorDeutsch.aslx to English.aslx and Deutsch.aslx, to avoid issues with gamebooks.

---
Removed JS.showScrollback(), as the jQuery-UI pop-ups cause errors with saved games and walkthroughs

Also removed the JS.showPopup* functions for the same reason.

(JS.printScrollback() is still intact.)

---
Updated transcript error message templates.

Added three try limit to `transcript_on_cmd` if `game.questplatform` does not exist and `JS.whereAmI()` for some reason repeatedly fails.