Fix new lines in transcript so they are always Environment.NewLine
The transcript was using Unix line breaks because of the way I was converting the HTML to text.

This appears to (hopefully) be the final touch concerning the transcript, without finding an alternate way to set game.questplatform to "desktop" from Play.vb or whichever file could do that. (Note that I don't really have anyone else testing things I'm changing right now, but I 'believe' I've tested things fairly extensively at this point.)

I am about to delete this branch from GitHub and create a new branch with 1 commit for this instead of the 22 commits I have created during the process. I am only submitting this for the record's sake.