Fix the transcript issues and change file format to TXT
There is a game-crashing bug in the transcript code I submitted for Quest 5.8.0.

If playing online and the transcript is enabled, the game completely crashes (because `SaveTranscript` is only in desktop.js, and is therefore undefined in the web player).

If playing in the desktop and the transcript is enabled, all turn scripts stop working completely and if a walkthrough is being recorded, the recorded steps are the HTML being output to the game.

ALSO, there was a JS string variable named `transcriptString` which got larger and larger when adding the current text for the transcript every time text printed to the game. This obviously slowed things down. Plus, this text was also sent by `ASLEvent` to update `game.transcriptstring` each time text printed along with everything else.

Also, the updated `clearScreen()` function in playercore.js was set up to save the cleared text and hide it rather than deleting it. This also negatively effected game performance exponentially as play progressed. The JS boolean `saveClearedText` is now set to false by default.

Finally, if the VIEW TRANSCRIPT command was run while recording a walkthrough, the jQuery-UI popup would break the walkthrough during playback.

The in-game HTML transcript feature has now been removed. There is no VIEW TRANSCRIPT command anymore. There is no longer a `transcriptString` JS variable, nor is there a `game.transcriptstring` attribute.

The only way the transcript will work now is if the game is being played in the desktop player. It will save the transcript, and append it, every time text prints to the screen, to an TXT file in "Documents\Quest Transcripts", just as in 5.8.0 except it is now TXT rather than HTML.

I also fixed my error in PlayerHTML.vb to make the file save under the transcript name the player chooses when enabling the transcript.

I added a "fallback" of each function I added to Player\desktopplayer.js to WebPlayer\player.js. These functions do nothing and serve no purpose except to avoid possible errors due to being undefined.

Also added a bit of code in `HandleCommand` to add the player's command to the transcript if `game.echocommand` is `false`.

WalkthroughRunner.vb has been modified to ignore anything starting with "event:WhereAmI" to stop the walkthrough from causing odd behaviour.

Player.vb has been modified to omit the "WhereAmI" event when recording walkthrough steps.

 I also added code to have the walkthrough ignore any function starting with "NoEventFunc" while running or recording, just in case authors would like other calls to `ASLEvent` to be ignored by the walkthrough.

Updated the transcript doc page to reflect the changes.

I triple checked things every way I could imagine in the desktop build as well as using VS to test the web player and the mobile version, but I haven't had anyone else test it. So, test it out, please.

---
I almost forgot:

If `saveClearedText` is changed to `true`, it will save the cleared text in #divOutput rather than dumping it.

The `printScrollback()` JS function is still present. Calling it will use the player's device to print the everything in #divOutput (whether via a printer or saving it to PDF).

Again, the scroll back will not contain any cleared text unless `saveClearedText` is changed from `false` to `true`.

The scrollback JS functions are pretty much hidden, but an author could easily add code to allow their use in their game.

---
All apologies to everyone for the bugs I introduced in Quest 5.8.0. I believe I have now fixed them all.

---
See issue #1054

---
For reference, the commit where the transcript feature was introduced to Quest: https://github.com/textadventures/quest/commit/a665d25b8b9e41996a1cb4124c6c34abd7686b9d