Upgrade to localStorage transcripts for online players
Now the transcript will save to a TXT file in "Documents\Quest Transcripts\" when using the desktop player, or to localStorage when using the web player or mobile.

I added a Transcripts button the web player (next to the Save button) and to the mobile player in the More menu.

The test game is here:

TODO: CSS for the transcripts list window

TODO: Should the Transcripts button disappear when the user has no transcripts in localStorage?