notranscript => transcriptprohibited
Changed the game object's attribute name from `notranscript` to `transcriptprohibited` to make the code easier to understand.

Also changed the JS counterpart `noTranscript` to `transcriptProhibited`.

Added a checkbox to the Quest GUI, providing the option for the author to prohibit the player from enabling the transcript.

Fixed copy and paste errors in comments for new JS functions.

---
Much thanks to @Pertex for all the feedback!