Remove the button and the checkbox
The button is no good. The site just needs a page for people to visit. (I have the page coded already, but I'm sure it can be improved, especially the CSS.)

I think the checkbox in the GUI to prohibit the use of the transcript would:
A) Never be used
B) Just confuse new users

Plus, the attribute `game.transcriptprohibited` is defined (false by default), and an author can just change that under the game's attributes in the GUI if desired.