Set saveClearedText to false by default
`saveClearedText` should have been `false` by default to begin with.

When false, `clearScreen` dumps all the data it clears from `#divOutput`. When true, a class is used to hide the data and preserve it to be displayed using `showScrollback`.

This also changes the display settings of the pop-up dialog displayed by `showScrollback` to fullscreen, so things will work properly regardless of screen dimensions.

NOTE: The change to `clearScreen` accidentally fixed an issue with `getCurrentDiv` in the WebPlayer. This will 're-introduce' that issue, but I will create an Issue to log that.