Save command will use JS.saveGame online
When saving progress online, the text from `#divOutput` is not retained when using `request`.

In 5.8, we changed it to use the game object attribute set by `JS.whereAmi()`, which ended up being unused (due to issues when run from the init script). This script initially included a call to whereAmI if the attribute was not set, which was also problematic, and it devolved into its current state -- which does not work on its own.

This will not use JS to check the platform and will still use JS to save when online.